+++ src/main.cpp
@@ -302,9 +302,18 @@ int main(int argument_count, const char** arguments) {
     }
     ss << file_stream.rdbuf();
   } else {
+#if defined(_WIN32)
+    std::cout << "Error. Reading from stdin is not supported on Windows. "
+                 "Please provide a file."
+              << std::endl;
+
+    std::cout << args;
+    return EXIT_FAILURE;
+#else
     std::cout << "Reading from stdin..." << std::flush;
     ss << std::cin.rdbuf();
     stdin = freopen("/dev/tty", "r", stdin);
+#endif
   }
 
   JSON json;