+++ CHANGELOG.md
@@ -7,4 +7,5 @@ unreleased:
 - Add '-f' or '--fullscreen" options. It causes the JSON to be displayed on the
   alternate buffer, in fullscreen.
 - Use 'q' or 'escape' to quit.
+- Support Windows. Reading the JSON from stdin is not supported.
 - Argument parser improvement.